requirements/base.txt
aplpy==2.0.3
astropy==4.0
astroquery==0.4
bokeh==2.2.0
cloudpickle==1.5.0
dask[dataframe]==2.15.0
dill==0.3.1.1
django==3.0.7
django-crispy-forms==1.9.2
django-environ==0.4.5
djangorestframework-datatables==0.5.1
djangorestframework==3.11.0
jinja2==2.11.2
numpy==1.18.1
pandas==1.0.1
psutil==5.7.0
psycopg2==2.8.4
pyarrow==0.17.0
scipy==1.4.1
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlalchemy==1.3.11
vaex==3.0.0
